pyproject.toml
# Source of truth for dependencies (use Poetry: pyproject.toml + poetry.lock).
# data/requirements.txt is generated for production pip installs, e.g.:
#   poetry export -f requirements.txt --without-hashes -o data/requirements.txt
# `poetry export` needs the install-time plugin: `poetry self add poetry-plugin-export`
# (development setup and CI add this; see scripts/setup.ps1 and .github/workflows).
# When you change deps, update the lockfile and re-export. The pre-commit hook
# scripts/$check_poetry_to_requirements.py runs export when pyproject.toml and
# poetry.lock are committed together (see .pre-commit-config.yaml).

[project]
name = "Kinderspielstadt-Los-Ämmerles--LA-Server"
version = "0.5.0"
description = "The la-server manages and stores all data required to digitalize the Kinderspielstadt.  It provides a cental backend that allows camp participants to connect clients and digitally perform and manage their jobs within the simulation."
authors = [
    {name = "Dieter Wellerdiek",email  = "dieter@los-aemmerles.org"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.14,<4.0"

dependencies = [
    "flask (>=3.1.3,<4.0.0)",
    "flask-cors (>=6.0.2,<7.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "pyjwt (>=2.8.0,<3.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "cryptography (>=46.0.5,<48.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "python-stdnum (>=2.2,<3.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "waitress (>=3.0.2,<4.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pre-commit (>=4.5.1,<5.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-xdist (>=3.8.0,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "ruff (>=0.15.7,<0.16.0)",
]

[tool.poetry]
package-mode = false
